Fixed transition between Main & Search
Program was closing when clicking cancel on Search window.
- Added if statement to main to check for a selected invoice.

Clicking Delete in Main would throw an exception occasionally.
- Set the DataGrid to only select a single entry at a time.
- Set the SelectedItem to null after each add or delete click.
- Adjusted the delete function to look for a row out of range.

+++ PcPartsInvoiceSystem/Search/clsSearchLogic.cs
@@ -15,7 +15,7 @@ namespace PcPartsInvoiceSystem.Search
         /// <summary>
         /// int of invoice cost that is selected, if none selected value is -1
         /// </summary>
-        private int invoiceCost;
+        private double invoiceCost;
         /// <summary>
         /// int of invoice number that is selected, if none selected value is -1
         /// </summary>
@@ -92,7 +92,7 @@ namespace PcPartsInvoiceSystem.Search
         /// sets invoice cost
         /// </summary>
         /// <param name="invoiceCost"></param>
-        public void UpdateInvoiceCost(int invoiceCost) 
+        public void UpdateInvoiceCost(double invoiceCost) 
         {
             this.invoiceCost = invoiceCost;
         }
